feat: ini ai bundler
