+++ apps/ai/package.json
@@ -31,7 +31,7 @@
     "@lightdotso/templates": "workspace:*",
     "@lightdotso/ui": "workspace:*",
     "@lightdotso/utils": "workspace:*",
-    "@next/bundle-analyzer": "15.0.0-canary.110",
+    "@next/bundle-analyzer": "15.0.0-canary.111",
     "@radix-ui/react-dialog": "^1.1.1",
     "@radix-ui/react-icons": "^1.3.0",
     "@sentry/integrations": "^7.114.0",
@@ -44,11 +44,12 @@
     "geist": "^1.3.1",
     "lucide-react": "^0.427.0",
     "nanoid": "^5.0.7",
-    "next": "^14.2.3",
+    "next": "15.0.0-canary.111",
     "next-themes": "^0.3.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "19.0.0-rc-187dd6a7-20240806",
+    "react-dom": "19.0.0-rc-187dd6a7-20240806",
     "react-intersection-observer": "^9.13.0",
+    "react-is": "19.0.0-rc-187dd6a7-20240806",
     "react-markdown": "^8.0.7",
     "react-syntax-highlighter": "^15.5.0",
     "react-textarea-autosize": "^8.5.3",
