+++ apps/ai/src/components/chat-panel.tsx
@@ -151,7 +151,6 @@ export function ChatPanel({
                     // open={shareDialogOpen}
                     // onOpenChange={setShareDialogOpen}
                     onCopy={() => setShareDialogOpen(false)}
-                    // @ts-expect-error
                     shareChatAction={shareChat}
                     chat={{
                       id,
