+++ apps/app/package.json
@@ -57,7 +57,7 @@
     "@lightdotso/utils": "workspace:*",
     "@lightdotso/validators": "workspace:*",
     "@lightdotso/wagmi": "workspace:*",
-    "@next/bundle-analyzer": "15.0.0-canary.110",
+    "@next/bundle-analyzer": "15.0.0-canary.111",
     "@prisma/client": "^5.18.0",
     "@prisma/nextjs-monorepo-workaround-plugin": "^5.18.0",
     "@radix-ui/react-dropdown-menu": "^2.1.1",
@@ -88,7 +88,7 @@
     "lucide-react": "^0.427.0",
     "msw": "^2.3.5",
     "neverthrow": "^7.0.0",
-    "next": "15.0.0-canary.110",
+    "next": "15.0.0-canary.111",
     "next-themes": "^0.3.0",
     "permissionless": "^0.1.44",
     "postgres": "^3.4.4",
