+++ apps/blog/package.json
@@ -40,7 +40,7 @@
     "autoprefixer": "^10.4.20",
     "lucide-react": "^0.427.0",
     "motion": "^10.18.0",
-    "next": "15.0.0-canary.110",
+    "next": "15.0.0-canary.111",
     "next-themes": "^0.3.0",
     "react": "19.0.0-rc-187dd6a7-20240806",
     "react-dom": "19.0.0-rc-187dd6a7-20240806",
