+++ apps/docs/package.json
@@ -26,13 +26,13 @@
     "@content-collections/next": "^0.2.0",
     "@fumadocs/content-collections": "^1.1.0",
     "@lightdotso/templates": "workspace:*",
-    "@next/bundle-analyzer": "15.0.0-canary.110",
+    "@next/bundle-analyzer": "15.0.0-canary.111",
     "@sentry/integrations": "^7.114.0",
     "@sentry/nextjs": "^8.25.0",
     "fumadocs-core": "^13.1.0",
     "fumadocs-openapi": "^3.3.0",
     "fumadocs-ui": "^13.1.0",
-    "next": "15.0.0-canary.110",
+    "next": "15.0.0-canary.111",
     "next-themes": "^0.3.0",
     "nextra": "^2.13.4",
     "nextra-theme-docs": "^2.13.4",
