+++ apps/explorer/package.json
@@ -40,7 +40,7 @@
     "@lightdotso/ui": "workspace:*",
     "@lightdotso/utils": "workspace:*",
     "@lightdotso/validators": "workspace:*",
-    "@next/bundle-analyzer": "15.0.0-canary.110",
+    "@next/bundle-analyzer": "15.0.0-canary.111",
     "@prisma/client": "^5.18.0",
     "@radix-ui/react-icons": "^1.3.0",
     "@sentry-internal/feedback": "^8.25.0",
@@ -54,7 +54,7 @@
     "kysely-postgres-js": "^2.0.0",
     "lucide-react": "^0.427.0",
     "neverthrow": "^7.0.0",
-    "next": "15.0.0-canary.110",
+    "next": "15.0.0-canary.111",
     "postgres": "^3.4.4",
     "prisma": "^5.18.0",
     "react": "19.0.0-rc-187dd6a7-20240806",
