+++ apps/frames/package.json
@@ -23,12 +23,12 @@
     "@babel/core": "^7.25.2",
     "@coinbase/onchainkit": "^0.13.4",
     "@lightdotso/services": "workspace:*",
-    "@next/bundle-analyzer": "15.0.0-canary.110",
+    "@next/bundle-analyzer": "15.0.0-canary.111",
     "@sentry/integrations": "^7.114.0",
     "@sentry/nextjs": "^8.25.0",
     "@vercel/og": "^0.6.2",
     "critters": "^0.0.24",
-    "next": "15.0.0-canary.110",
+    "next": "15.0.0-canary.111",
     "react": "19.0.0-rc-187dd6a7-20240806",
     "react-dom": "19.0.0-rc-187dd6a7-20240806",
     "viem": "^2.13.7"
