+++ apps/home/package.json
@@ -27,14 +27,14 @@
     "@lightdotso/templates": "workspace:*",
     "@lightdotso/ui": "workspace:*",
     "@lightdotso/utils": "workspace:*",
-    "@next/bundle-analyzer": "15.0.0-canary.110",
+    "@next/bundle-analyzer": "15.0.0-canary.111",
     "@radix-ui/react-icons": "^1.3.0",
     "@sentry/integrations": "^7.114.0",
     "@sentry/nextjs": "^8.25.0",
     "critters": "^0.0.24",
     "framer-motion": "^11.3.24",
     "lucide-react": "^0.427.0",
-    "next": "15.0.0-canary.110",
+    "next": "15.0.0-canary.111",
     "react": "19.0.0-rc-187dd6a7-20240806",
     "react-dom": "19.0.0-rc-187dd6a7-20240806",
     "viem": "^2.19.3"
