+++ apps/inngest/package.json
@@ -26,7 +26,7 @@
     "@babel/core": "^7.25.2",
     "@lightdotso/prisma": "workspace:*",
     "@lightdotso/redis": "workspace:*",
-    "@next/bundle-analyzer": "15.0.0-canary.110",
+    "@next/bundle-analyzer": "15.0.0-canary.111",
     "@prisma/nextjs-monorepo-workaround-plugin": "^5.18.0",
     "@sentry-internal/feedback": "^8.25.0",
     "@sentry/integrations": "^7.114.0",
@@ -39,7 +39,7 @@
     "@upstash/redis": "^1.34.0",
     "critters": "^0.0.24",
     "inngest": "^3.22.2",
-    "next": "15.0.0-canary.110",
+    "next": "15.0.0-canary.111",
     "react": "19.0.0-rc-187dd6a7-20240806",
     "react-dom": "19.0.0-rc-187dd6a7-20240806",
     "zod": "^3.23.8"
