+++ packages/dialogs/package.json
@@ -43,7 +43,7 @@
     "@lightdotso/wagmi": "workspace:*",
     "class-variance-authority": "^0.7.0",
     "lucide-react": "^0.427.0",
-    "next": "15.0.0-canary.110",
+    "next": "15.0.0-canary.111",
     "permissionless": "^0.1.44",
     "react": "19.0.0-rc-187dd6a7-20240806",
     "react-dom": "19.0.0-rc-187dd6a7-20240806",
