+++ packages/elements/package.json
@@ -33,7 +33,7 @@
     "@radix-ui/react-icons": "^1.3.0",
     "class-variance-authority": "^0.7.0",
     "lucide-react": "^0.427.0",
-    "next": "15.0.0-canary.110",
+    "next": "15.0.0-canary.111",
     "react": "19.0.0-rc-187dd6a7-20240806",
     "react-blurhash": "^0.3.0",
     "react-dom": "19.0.0-rc-187dd6a7-20240806",
