+++ packages/hooks/package.json
@@ -37,7 +37,7 @@
     "copy-to-clipboard": "^3.3.3",
     "lodash": "^4.17.21",
     "merkletreejs": "^0.4.0",
-    "next": "15.0.0-canary.110",
+    "next": "15.0.0-canary.111",
     "react": "19.0.0-rc-187dd6a7-20240806",
     "react-dom": "19.0.0-rc-187dd6a7-20240806",
     "react-is": "19.0.0-rc-187dd6a7-20240806",
