+++ packages/popup/package.json
@@ -31,7 +31,7 @@
     "dotenv": "^16.4.5",
     "events": "^3.3.0",
     "isutf8": "^4.0.0",
-    "next": "15.0.0-canary.110",
+    "next": "15.0.0-canary.111",
     "react": "19.0.0-rc-187dd6a7-20240806",
     "react-dom": "19.0.0-rc-187dd6a7-20240806",
     "react-is": "19.0.0-rc-187dd6a7-20240806"
