+++ packages/tables/package.json
@@ -41,7 +41,7 @@
     "@tanstack/react-table": "^8.20.1",
     "@tremor/react": "^3.17.4",
     "lucide-react": "^0.427.0",
-    "next": "15.0.0-canary.110",
+    "next": "15.0.0-canary.111",
     "react": "19.0.0-rc-187dd6a7-20240806",
     "react-dom": "19.0.0-rc-187dd6a7-20240806",
     "react-is": "19.0.0-rc-187dd6a7-20240806",
