+++ packages/templates/package.json
@@ -44,7 +44,7 @@
     "class-variance-authority": "^0.7.0",
     "framer-motion": "^11.3.24",
     "lucide-react": "^0.427.0",
-    "next": "15.0.0-canary.110",
+    "next": "15.0.0-canary.111",
     "next-themes": "^0.3.0",
     "react": "19.0.0-rc-187dd6a7-20240806",
     "react-dom": "19.0.0-rc-187dd6a7-20240806",
