+++ packages/ui/package.json
@@ -74,7 +74,7 @@
     "embla-carousel-react": "^8.1.3",
     "framer-motion": "^11.3.24",
     "lucide-react": "^0.263.1",
-    "next": "15.0.0-canary.110",
+    "next": "15.0.0-canary.111",
     "next-themes": "^0.3.0",
     "react": "19.0.0-rc-187dd6a7-20240806",
     "react-blurhash": "^0.3.0",
